Prestige Forest Hills Mulund West Mumbai
Prestige Forest Hills is a New Residential Project 2, 3 & 4 BHK Luxury Apartments located at Mulund West Mumbai. The project offers an extensive assortment of facilities and the area boasts of good connectivity with other main areas. Prestige Mulund Mumbai apartments are expansive enough for a new and contemporary family to set up a home and because of the wide range of facilities available in the vicinity; one can enjoy a premium lifestyle. 

As one of India's most revealed names in Real Estate, Prestige Group Real Estate takes many years of Inheritance towards the progress of Prestige Forest Hills Mulund, sincerely splendid accountability that sets the norm in luxury living. This pointless high rise is set to be one of the utmost in India and is a great idea to go to overpower the city's viewpoint. The trust and certainty that we have gathered from our extras and customers as the years progressed, is contributory to assuring that we scale substantially more well-known heights with this venture and take top-notch living preparations that proceed over the rest.
 
These are the hometowns that suit your personality. The abode is expansive and makes the whole living involvement valuable and impressive. The residence is carefully intended and will give you all the honor and pride that you truly deserve. For a new start in life, this is an innovative place prepared with all the modern facilities required for a life of comfort and style. Prestige Forest Hills Mulund Mumbai promises you to give a rich and plentiful lifestyle, with great ease of accessibility to all kinds of facilities. One of the leading ventures by Prestige Group, which integrates all the basics for healthy growth for you and your family.

The Residential Apartments is positioned at Mulund West, to make your mornings and life lovely. New Launch Prestige Forest Hills is surrounded by a luxurious green atmosphere and encases the landscape view of yogi hills and the stunning panoramic cityscape of Mumbai. It gives an Apartment complex loaded with all facilities and features.

Mulund West, Mumbai location

Mulund is an area in the northeast of Mumbai, Maharashtra, India. It is also a railway station on the Central Station of the Mumbai Residential Railway. It is also the last railway station of the Mumbai Suburban district on the vital railway line.  It is nestled together with the foothills of the Sanjay Gandhi National Park with easy access to the Eastern Express Highway and Navi Mumbai finished the Mulund-Airoli Bridge. Mulund marks the end of Mumbai city Mulund is a very posh housing place. It enjoys brilliant connectivity with Eastern Express Highway, Western Express Highway, Thane, Whole Navi Mumbai, and Rest of the Mumbai. It is well linked by road and rail, transport with Navi Mumbai, Western Suburbs, and Thane.

For More Details :
Visit - https://prestige.developerprojects.in/mumbai/prestige-forest-hills-mulund/